server/requirements.txt
fastapi
uvicorn
gspread
google-auth
google-auth-oauthlib
python-dotenv
pytest
